always include `setup` system, even when not selecting a camera Fixes #9
